fixed crash when dumping unexpected data
This actually happened when trying to dump the params of a builder.apply event in TPF2 when building a street cargo station (large one). Feel free to investigate the actual reason. 

PS: L63 sortKeys is an undefined global. Didn't understand the intention, so I have no fix. But maybe it's related.